Clean up egg-herbie code
